to bfloat16

+++ owl_wms/trainers/world_trainer.py
@@ -216,7 +216,7 @@ class WorldTrainer(BaseTrainer):
 
     def prep_batch(self, batch):
         """Move to cuda, and if necessary use encoder to convert rgb to latent (x)"""
-        batch = {k: v.cuda() if isinstance(v, torch.Tensor) else v for k, v in batch.items()}
+        batch = {k: v.cuda().bfloat16() if isinstance(v, torch.Tensor) else v for k, v in batch.items()}
         if "rgb" in batch:
             assert "x" not in batch, "passed rgb to convert, but already have batch item `x` (latents)"
             batch["x"] = self.encoder_decoder.encode(batch.pop("rgb")).bfloat16()
